Move hash-related logic to a separate class
This is a preparation for adding unit tests for hash comparison logic.
